refactored api calls out
