+++ .gitignore
@@ -5,5 +5,6 @@ test/
 Gooey/
 *.mbox
 *.csv
+*.pyc
 .env
 .DS_Store
